+++ nextjs-crud/.eslintrc.json
