+++ nextjs-crud/app/components/edit-product.js
@@ -1,47 +1,64 @@
 "use client";
-import React from "react";
-import BreadCrumb from "@/app/components/bread-crumb";
+import React, { useState, useEffect } from "react";
+import BreadCrumb from "@/nextjs-crud/app/components/bread-crumb";
 import { useForm } from "react-hook-form";
 import { useRouter } from "next/navigation";
 
 const breadCrumb = [
   { title: "Home", url: "../" },
-  { title: "Add New Product", url: "../add/" },
+  { title: "Edit Product", url: "../edit/" },
 ];
 
-const AddProduct = () => {
+const EditProduct = ({ id }) => {
   const router = useRouter();
-
   const {
     register,
     handleSubmit,
     formState: { errors },
-  } = useForm();
+  } = useForm({
+    defaultValues: async () => {
+      const { product } = await getProduct(id);
+      return product;
+    },
+  });
 
   const onSubmit = async (data) => {
     try {
-      const res = await fetch("../api", {
+      const res = await fetch(`../api/${id}/`, {
         method: "POST",
         body: JSON.stringify(data),
       });
 
       if (!res.ok) {
-        throw new Error("Failed to add product");
+        throw new Error("Failed to update product");
       }
 
       const { message } = await res.json();
       alert(message);
       router.push("../");
     } catch (error) {
-      console.log("Failed to add product", error);
-      alert("Failed to add product");
+      console.log("Failed to update product", error);
+      alert("Failed to update product");
+    }
+  };
+
+  const getProduct = async (id) => {
+    try {
+      const res = await fetch(`../api/${id}`);
+      if (!res.ok) {
+        throw new Error("Failed to get product");
+      }
+
+      return await res.json();
+    } catch (error) {
+      alert("Failed to get product");
     }
   };
 
   return (
     <div>
       <BreadCrumb lists={breadCrumb} />
-      <h4 className="mb-2">Add New Product</h4>
+      <h4 className="mb-2">Edit Product</h4>
       <div className="mb-2">
         <div className="row">
           <div className="col-md-6">
@@ -52,7 +69,9 @@ const AddProduct = () => {
                 </label>
                 <input
                   className="form-control"
-                  {...register("title", { required: true })}
+                  {...register("title", {
+                    required: true,
+                  })}
                 />
               </div>
               <div className="mb-3">
@@ -76,6 +95,8 @@ const AddProduct = () => {
               <div className="mb-3 text-end">
                 <input type="submit" className="btn btn-primary" />
               </div>
+
+              {errors.exampleRequired && <span>This field is required</span>}
             </form>
           </div>
         </div>
@@ -84,4 +105,4 @@ const AddProduct = () => {
   );
 };
 
-export default AddProduct;
+export default EditProduct;
