+++ nextjs-crud/app/components/view-product.js
@@ -1,46 +1,21 @@
 "use client";
 import React, { useState, useEffect } from "react";
-import BreadCrumb from "@/app/components/bread-crumb";
+import BreadCrumb from "@/nextjs-crud/app/components/bread-crumb";
 import { useForm } from "react-hook-form";
 import { useRouter } from "next/navigation";
 
 const breadCrumb = [
   { title: "Home", url: "../" },
-  { title: "Edit Product", url: "../edit/" },
+  { title: "View Product", url: "../view/" },
 ];
 
-const EditProduct = ({ id }) => {
-  const router = useRouter();
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
+const ViewProduct = ({ id }) => {
+  const { register } = useForm({
     defaultValues: async () => {
       const { product } = await getProduct(id);
       return product;
     },
   });
-
-  const onSubmit = async (data) => {
-    try {
-      const res = await fetch(`../api/${id}/`, {
-        method: "POST",
-        body: JSON.stringify(data),
-      });
-
-      if (!res.ok) {
-        throw new Error("Failed to update product");
-      }
-
-      const { message } = await res.json();
-      alert(message);
-      router.push("../");
-    } catch (error) {
-      console.log("Failed to update product", error);
-      alert("Failed to update product");
-    }
-  };
 
   const getProduct = async (id) => {
     try {
@@ -58,11 +33,11 @@ const EditProduct = ({ id }) => {
   return (
     <div>
       <BreadCrumb lists={breadCrumb} />
-      <h4 className="mb-2">Edit Product</h4>
+      <h4 className="mb-2">View Product</h4>
       <div className="mb-2">
         <div className="row">
           <div className="col-md-6">
-            <form onSubmit={handleSubmit(onSubmit)} method="POST">
+            <form method="POST">
               <div className="mb-3">
                 <label htmlFor="title" className="form-label">
                   Title
@@ -70,7 +45,7 @@ const EditProduct = ({ id }) => {
                 <input
                   className="form-control"
                   {...register("title", {
-                    required: true,
+                    disabled: true,
                   })}
                 />
               </div>
@@ -80,7 +55,7 @@ const EditProduct = ({ id }) => {
                 </label>
                 <textarea
                   className="form-control"
-                  {...register("description", { required: true })}
+                  {...register("description", { disabled: true })}
                 ></textarea>
               </div>
               <div className="mb-3">
@@ -89,14 +64,9 @@ const EditProduct = ({ id }) => {
                 </label>
                 <input
                   className="form-control"
-                  {...register("price", { required: true })}
+                  {...register("price", { disabled: true })}
                 />
               </div>
-              <div className="mb-3 text-end">
-                <input type="submit" className="btn btn-primary" />
-              </div>
-
-              {errors.exampleRequired && <span>This field is required</span>}
             </form>
           </div>
         </div>
@@ -105,4 +75,4 @@ const EditProduct = ({ id }) => {
   );
 };
 
-export default EditProduct;
+export default ViewProduct;
