+++ nextjs-crud/next.config.js
